move image utilities to another file
This is to use the camera class only to capture images.
